+++ packages/client/modules/demo/initDB.ts
@@ -1,7 +1,7 @@
 import {SlackNotificationEventEnum} from '~/__generated__/SlackNotificationList_viewer.graphql'
 import {PALETTE} from '~/styles/paletteV3'
-import RetrospectiveMeeting from '../../../server/database/types/MeetingRetrospective'
 import ITask from '../../../server/database/types/Task'
+import {RetrospectiveMeeting} from '../../../server/postgres/types/Meeting'
 import JiraProjectId from '../../shared/gqlIds/JiraProjectId'
 import demoUserAvatar from '../../styles/theme/images/avatar-user.svg'
 import {ExternalLinks, MeetingSettingsThreshold, RetroDemo} from '../../types/constEnums'