+++ packages/client/mutations/LoginWithGoogleMutation.ts
@@ -55,7 +55,7 @@ const LoginWithGoogleMutation: StandardMutation<TLoginWithGoogleMutation, Histor
       handleAcceptTeamInvitationErrors(atmosphere, acceptTeamInvitation)
       if (!uiError && !errors) {
         handleSuccessfulLogin(loginWithGoogle)
-        const authToken = acceptTeamInvitation?.authToken ?? loginWithGoogle.authToken
+        const authToken = acceptTeamInvitation?.authToken || loginWithGoogle.authToken!
         atmosphere.setAuthToken(authToken)
         const redirectPath =
           isNewUser && user?.hasSignUpDestinationTeamFlag