+++ packages/client/utils/meetings/lookups.ts
@@ -1,6 +1,6 @@
 import React from 'react'
-import {MeetingTypeEnum} from '~/../server/postgres/types/Meeting'
 import {NewMeetingPhaseTypeEnum} from '~/__generated__/ActionMeetingSidebar_meeting.graphql'
+import {MeetingTypeEnum} from '../../__generated__/SummarySheet_meeting.graphql'
 import CardsSVG from '../../components/CardsSVG'
 import {ACTION, POKER, RETROSPECTIVE, TEAM_PROMPT} from '../constants'