+++ packages/server/database/types/MeetingAction.ts
@@ -1,10 +1,6 @@
-import AgendaItemsPhase from './AgendaItemsPhase'
-import CheckInPhase from './CheckInPhase'
-import GenericMeetingPhase from './GenericMeetingPhase'
+import {CheckInMeetingPhase} from '../../postgres/types/NewMeetingPhase'
 import Meeting from './Meeting'
-import UpdatesPhase from './UpdatesPhase'
 
-type CheckInMeetingPhase = CheckInPhase | UpdatesPhase | GenericMeetingPhase | AgendaItemsPhase
 interface Input {
   id?: string
   teamId: string
@@ -12,10 +8,6 @@ interface Input {
   name: string
   phases: [CheckInMeetingPhase, ...CheckInMeetingPhase[]]
   facilitatorUserId: string
-}
-
-export function isMeetingAction(meeting: Meeting): meeting is MeetingAction {
-  return meeting.meetingType === 'action'
 }
 
 export default class MeetingAction extends Meeting {