+++ packages/server/database/types/MeetingRetrospective.ts
@@ -1,44 +1,31 @@
-import GenericMeetingPhase from './GenericMeetingPhase'
+import {AutogroupReflectionGroupType, TranscriptBlock} from '../../postgres/types'
+import {RetroMeetingPhase} from '../../postgres/types/NewMeetingPhase'
 import Meeting from './Meeting'
 
-export type AutogroupReflectionGroupType = {
-  groupTitle: string
-  reflectionIds: string[]
-}
-
-export type TranscriptBlock = {
-  speaker: string
-  words: string
-}
-
 interface Input {
-  id?: string
+  id?: string | null
   teamId: string
   meetingCount: number
   name: string
-  phases: [GenericMeetingPhase, ...GenericMeetingPhase[]]
+  phases: [RetroMeetingPhase, ...RetroMeetingPhase[]]
   facilitatorUserId: string
-  showConversionModal?: boolean
+  showConversionModal?: boolean | null
   templateId: string
   totalVotes: number
   maxVotesPerGroup: number
   disableAnonymity: boolean
-  transcription?: TranscriptBlock[]
-  autogroupReflectionGroups?: AutogroupReflectionGroupType[]
-  resetReflectionGroups?: AutogroupReflectionGroupType[]
+  transcription?: TranscriptBlock[] | null
+  autogroupReflectionGroups?: AutogroupReflectionGroupType[] | null
+  resetReflectionGroups?: AutogroupReflectionGroupType[] | null
   recallBotId?: string
-  videoMeetingURL?: string
-  meetingSeriesId?: number
+  videoMeetingURL?: string | null
+  meetingSeriesId?: number | null
   scheduledEndTime?: Date | null
-}
-
-export function isMeetingRetrospective(meeting: Meeting): meeting is MeetingRetrospective {
-  return meeting.meetingType === 'retrospective'
 }
 
 export default class MeetingRetrospective extends Meeting {
   meetingType!: 'retrospective'
-  showConversionModal?: boolean
+  showConversionModal?: boolean | null
   autoGroupThreshold?: number | null
   nextAutoGroupThreshold?: number | null
   totalVotes: number
@@ -50,11 +37,11 @@ export default class MeetingRetrospective extends Meeting {
   templateId: string
   topicCount?: number
   reflectionCount?: number
-  transcription?: TranscriptBlock[]
-  recallBotId?: string
-  videoMeetingURL?: string
-  autogroupReflectionGroups?: AutogroupReflectionGroupType[]
-  resetReflectionGroups?: AutogroupReflectionGroupType[]
+  transcription?: TranscriptBlock[] | null
+  recallBotId?: string | null
+  videoMeetingURL?: string | null
+  autogroupReflectionGroups?: AutogroupReflectionGroupType[] | null
+  resetReflectionGroups?: AutogroupReflectionGroupType[] | null
 
   constructor(input: Input) {
     const {