+++ packages/server/graphql/mutations/helpers/generateGroups.ts
@@ -1,7 +1,7 @@
 import {SubscriptionChannel} from '../../../../client/types/constEnums'
 import getRethink from '../../../database/rethinkDriver'
-import {AutogroupReflectionGroupType} from '../../../database/types/MeetingRetrospective'
-import {RetroReflection} from '../../../postgres/types'
+import getKysely from '../../../postgres/getKysely'
+import {AutogroupReflectionGroupType, RetroReflection} from '../../../postgres/types'
 import {Logger} from '../../../utils/Logger'
 import OpenAIServerManager from '../../../utils/OpenAIServerManager'
 import {analytics} from '../../../utils/analytics/analytics'
@@ -54,6 +54,11 @@ const generateGroups = async (
   }
 
   const r = await getRethink()
+  await getKysely()
+    .updateTable('NewMeeting')
+    .set({autogroupReflectionGroups: JSON.stringify(autogroupReflectionGroups)})
+    .where('id', '=', meetingId)
+    .execute()
   const meetingRes = await r
     .table('NewMeeting')
     .get(meetingId)