+++ packages/server/graphql/mutations/helpers/safeEndRetrospective.ts
@@ -1,12 +1,13 @@
+import {sql} from 'kysely'
 import {SubscriptionChannel} from 'parabol-client/types/constEnums'
 import {DISCUSS} from 'parabol-client/utils/constants'
 import getMeetingPhase from 'parabol-client/utils/getMeetingPhase'
 import findStageById from 'parabol-client/utils/meetings/findStageById'
 import {checkTeamsLimit} from '../../../billing/helpers/teamLimitsCheck'
 import getRethink from '../../../database/rethinkDriver'
-import MeetingRetrospective from '../../../database/types/MeetingRetrospective'
 import TimelineEventRetroComplete from '../../../database/types/TimelineEventRetroComplete'
 import getKysely from '../../../postgres/getKysely'
+import {RetrospectiveMeeting} from '../../../postgres/types/Meeting'
 import removeSuggestedAction from '../../../safeMutations/removeSuggestedAction'
 import {Logger} from '../../../utils/Logger'
 import RecallAIServerManager from '../../../utils/RecallAIServerManager'
@@ -33,14 +34,15 @@ const getTranscription = async (recallBotId?: string | null) => {
   return await manager.getBotTranscript(recallBotId)
 }
 
-const summarizeRetroMeeting = async (meeting: MeetingRetrospective, context: InternalContext) => {
+const summarizeRetroMeeting = async (meeting: RetrospectiveMeeting, context: InternalContext) => {
   const {dataLoader} = context
   const {id: meetingId, phases, facilitatorUserId, teamId, recallBotId} = meeting
+  const pg = getKysely()
   const r = await getRethink()
   const [reflectionGroups, reflections, sentimentScore] = await Promise.all([
     dataLoader.get('retroReflectionGroupsByMeetingId').load(meetingId),
     dataLoader.get('retroReflectionsByMeetingId').load(meetingId),
-    generateWholeMeetingSentimentScore(meetingId, facilitatorUserId, dataLoader)
+    generateWholeMeetingSentimentScore(meetingId, facilitatorUserId!, dataLoader)
   ])
   const discussPhase = getPhase(phases, 'discuss')
   const {stages} = discussPhase
@@ -48,24 +50,39 @@ const summarizeRetroMeeting = async (meeting: MeetingRetrospective, context: Int
 
   const reflectionGroupIds = reflectionGroups.map(({id}) => id)
   const [summary, transcription] = await Promise.all([
-    generateWholeMeetingSummary(discussionIds, meetingId, teamId, facilitatorUserId, dataLoader),
+    generateWholeMeetingSummary(discussionIds, meetingId, teamId, facilitatorUserId!, dataLoader),
     getTranscription(recallBotId)
   ])
   const commentCounts = (
     await dataLoader.get('commentCountByDiscussionId').loadMany(discussionIds)
   ).filter(isValid)
   const commentCount = commentCounts.reduce((cumSum, count) => cumSum + count, 0)
+  const taskCount = await r
+    .table('Task')
+    .getAll(r.args(discussionIds), {index: 'discussionId'})
+    .count()
+    .default(0)
+    .run()
+  await pg
+    .updateTable('NewMeeting')
+    .set({
+      commentCount,
+      taskCount,
+      topicCount: reflectionGroupIds.length,
+      reflectionCount: reflections.length,
+      sentimentScore,
+      summary,
+      transcription
+    })
+    .where('id', '=', meetingId)
+    .execute()
   await r
     .table('NewMeeting')
     .get(meetingId)
     .update(
       {
         commentCount,
-        taskCount: r
-          .table('Task')
-          .getAll(r.args(discussionIds), {index: 'discussionId'})
-          .count()
-          .default(0) as unknown as number,
+        taskCount,
         topicCount: reflectionGroupIds.length,
         reflectionCount: reflections.length,
         sentimentScore,
@@ -76,7 +93,7 @@ const summarizeRetroMeeting = async (meeting: MeetingRetrospective, context: Int
     )
     .run()
 
-  dataLoader.get('newMeetings').clear(meetingId)
+  dataLoader.clearAll('newMeetings')
   // wait for whole meeting summary to be generated before sending summary email and updating qualAIMeetingCount
   sendNewMeetingSummary(meeting, context).catch(Logger.log)
   updateQualAIMeetingsCount(meetingId, teamId, dataLoader)
@@ -93,7 +110,7 @@ const safeEndRetrospective = async ({
   context,
   now
 }: {
-  meeting: MeetingRetrospective
+  meeting: RetrospectiveMeeting
   context: InternalContext
   now: Date
 }) => {
@@ -115,7 +132,7 @@ const safeEndRetrospective = async ({
   const phase = getMeetingPhase(phases)
 
   const insights = await gatherInsights(meeting, dataLoader)
-  const completedRetrospective = (await r
+  const completedRetrospective = await r
     .table('NewMeeting')
     .get(meetingId)
     .update(
@@ -127,21 +144,35 @@ const safeEndRetrospective = async ({
       {returnChanges: true}
     )('changes')(0)('new_val')
     .default(null)
-    .run()) as unknown as MeetingRetrospective
-
+    .run()
+  await getKysely()
+    .updateTable('NewMeeting')
+    .set({
+      endedAt: sql`CURRENT_TIMESTAMP`,
+      phases: JSON.stringify(phases),
+      usedReactjis: JSON.stringify(insights.usedReactjis),
+      engagement: insights.engagement
+    })
+    .where('id', '=', meetingId)
+    .executeTakeFirst()
+  dataLoader.clearAll('newMeetings')
   if (!completedRetrospective) {
     return standardError(new Error('Completed retrospective meeting does not exist'), {
       userId: viewerId
     })
   }
-
+  if (completedRetrospective.meetingType !== 'retrospective') {
+    return standardError(new Error('Meeting type is not retrospective'), {
+      userId: viewerId
+    })
+  }
   // remove any empty tasks
   const {templateId} = completedRetrospective
   const [meetingMembers, team, teamMembers, removedTaskIds, template] = await Promise.all([
     dataLoader.get('meetingMembersByMeetingId').load(meetingId),
     dataLoader.get('teams').loadNonNull(teamId),
     dataLoader.get('teamMembersByTeamId').load(teamId),
-    removeEmptyTasks(meetingId),
+    removeEmptyTasks(meetingId, teamId),
     dataLoader.get('meetingTemplates').loadNonNull(templateId),
     updateTeamInsights(teamId, dataLoader)
   ])