+++ packages/server/postgres/select.ts
@@ -2,8 +2,8 @@ import type {JSONContent} from '@tiptap/core'
 import {NotNull, sql} from 'kysely'
 import {NewMeetingPhaseTypeEnum} from '../graphql/public/resolverTypes'
 import getKysely from './getKysely'
-import {ReactjiDB} from './types'
-
+import {AutogroupReflectionGroupType, ReactjiDB, TranscriptBlock, UsedReactjis} from './types'
+import type {NewMeetingPhase} from './types/NewMeetingPhase'
 export const selectTimelineEvent = () => {
   return getKysely().selectFrom('TimelineEvent').selectAll().$narrowType<
     | {
@@ -229,3 +229,53 @@ export const selectComments = () =>
     .select(({fn}) => [fn<ReactjiDB[]>('to_json', ['reactjis']).as('reactjis')])
 
 export const selectReflectPrompts = () => getKysely().selectFrom('ReflectPrompt').selectAll()
+
+export const selectNewMeetings = () =>
+  getKysely()
+    .selectFrom('NewMeeting')
+    .select(({fn}) => [
+      'id',
+      'isLegacy',
+      'createdAt',
+      'updatedAt',
+      'createdBy',
+      'endedAt',
+      'facilitatorStageId',
+      'facilitatorUserId',
+      'meetingCount',
+      'meetingNumber',
+      'name',
+      'summarySentAt',
+      'teamId',
+      'meetingType',
+      'showConversionModal',
+      'meetingSeriesId',
+      'scheduledEndTime',
+      'summary',
+      'sentimentScore',
+      'slackTs',
+      'engagement',
+      'totalVotes',
+      'maxVotesPerGroup',
+      'disableAnonymity',
+      'commentCount',
+      'taskCount',
+      'agendaItemCount',
+      'storyCount',
+      'templateId',
+      'topicCount',
+      'reflectionCount',
+      'recallBotId',
+      'videoMeetingURL',
+      'templateRefId',
+      'meetingPrompt',
+      fn<NewMeetingPhase[]>('to_json', ['phases']).as('phases'),
+      fn<UsedReactjis | null>('to_json', ['usedReactjis']).as('usedReactjis'),
+      fn<TranscriptBlock[] | null>('to_json', ['transcription']).as('transcription'),
+      fn<AutogroupReflectionGroupType[] | null>('to_json', ['autogroupReflectionGroups']).as(
+        'autogroupReflectionGroups'
+      ),
+      fn<AutogroupReflectionGroupType[] | null>('to_json', ['resetReflectionGroups']).as(
+        'resetReflectionGroups'
+      )
+    ])